# docs/api-ledger/openapi-snapshot.json
{
  "components" : {
    "schemas" : {
...
      "CreateFuelRequestResource" : {
        "properties" : {
          "buyerCompanyId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "deliveryAddress" : {
            "type" : "string"
          },
          "deliveryDate" : {
            "format" : "date",
            "type" : "string"
          },
          "equipmentId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "fuelProductId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "providerId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "quantity" : {
            "format" : "double",
            "type" : "number"
          },
          "source" : {
            "type" : "string"
          },
          "unit" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "FuelRequestResource" : {
        "properties" : {
          "buyerCompanyId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "deliveryAddress" : {
            "type" : "string"
          },
          "deliveryDate" : {
            "format" : "date",
            "type" : "string"
          },
          "equipmentId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "fuelProductId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "fuelType" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "productName" : {
            "type" : "string"
          },
          "providerId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "quantity" : {
            "format" : "double",
            "type" : "number"
          },
          "rejectionReason" : {
            "type" : "string"
          },
          "source" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING", "APPROVED", "REJECTED" ],
            "type" : "string"
          },
          "unit" : {
            "type" : "string"
          },
          "unitPrice" : {
            "format" : "double",
            "type" : "number"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RejectFuelRequestResource" : {
        "properties" : {
          "reason" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Token JWT Bearer para autenticar solicitudes a la API",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "support@primefuel.com",
      "name" : "Soporte de FullTank",
      "url" : "https://primefuel.com/support"
    },
    "description" : "API de FullTank para gestionar abastecimiento, flota, entregas, inventario y pagos por organizaciÃ³n.",
    "license" : {
      "name" : "Apache 2.0 (licencia)",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "FullTank Platform",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/fuel-requests" : {
      "get" : {
        "description" : "Devuelve solicitudes de una sola dimensión propia: la empresa compradora o el tenant distribuidor del usuario.",
        "operationId" : "findAll",
        "parameters" : [ {
          "in" : "query",
          "name" : "buyerCompanyId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "providerId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/FuelRequestResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Se devuelve la lista de solicitudes."
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/FuelRequestResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Se indicaron a la vez buyerCompanyId y providerId."
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/FuelRequestResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "No se indicó filtro o el filtro indicado no pertenece al usuario."
          }
        },
        "summary" : "Listar solicitudes de combustible",
        "tags" : [ "Solicitudes de combustible heredadas" ]
      },
      "post" : {
        "description" : "Crea una solicitud v1 y su solicitud de reposición vinculada para una empresa compradora del usuario autenticado.",
        "operationId" : "create_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateFuelRequestResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FuelRequestResource"
                }
              }
            },
            "description" : "Solicitud de combustible creada."
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FuelRequestResource"
                }
              }
            },
            "description" : "El producto de combustible no existe o no pertenece al distribuidor indicado."
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FuelRequestResource"
                }
              }
            },
            "description" : "La empresa compradora del cuerpo no pertenece al usuario autenticado."
          }
        },
        "summary" : "Crear solicitud de combustible heredada",
        "tags" : [ "Solicitudes de combustible heredadas" ]
      }
    },
    "/api/v1/fuel-requests/{requestId}" : {
      "get" : {
        "description" : "Devuelve la solicitud indicada cuando el usuario pertenece a la empresa compradora o al distribuidor correspondiente.",
        "operationId" : "findById",
        "parameters" : [ {
          "in" : "path",
          "name" : "requestId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FuelRequestResource"
                }
              }
            },
            "description" : "Solicitud de combustible devuelta."
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FuelRequestResource"
                }
              }
            },
            "description" : "La solicitud no existe o no pertenece al usuario autenticado."
          }
        },
        "summary" : "Consultar solicitud de combustible por identificador",
        "tags" : [ "Solicitudes de combustible heredadas" ]
      }
    },
    "/api/v1/fuel-requests/{requestId}/accept" : {
      "post" : {
        "description" : "Acepta una solicitud pendiente en nombre del distribuidor destinatario y crea la orden de combustible resultante.",
        "operationId" : "accept_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "requestId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Solicitud aceptada y orden de combustible creada."
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "No se encontró la solicitud o su producto al crear la orden."
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "La solicitud no existe o no está dirigida al tenant distribuidor autenticado."
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "La solicitud ya fue aceptada o no se pudo aceptar la revisión vinculada."
          }
        },
        "summary" : "Aceptar solicitud de combustible",
        "tags" : [ "Solicitudes de combustible heredadas" ]
      }
    },
    "/api/v1/fuel-requests/{requestId}/reject" : {
      "post" : {
        "description" : "Rechaza una solicitud pendiente para el distribuidor destinatario y registra el motivo indicado.",
        "operationId" : "reject_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "requestId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RejectFuelRequestResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FuelRequestResource"
                }
              }
            },
            "description" : "Solicitud de combustible rechazada."
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FuelRequestResource"
                }
              }
            },
            "description" : "Falta el motivo de rechazo o no se encontró la solicitud."
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FuelRequestResource"
                }
              }
            },
            "description" : "La solicitud no existe o no está dirigida al tenant distribuidor autenticado."
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FuelRequestResource"
                }
              }
            },
            "description" : "La solicitud no está pendiente y no puede rechazarse."
          }
        },
        "summary" : "Rechazar solicitud de combustible",
        "tags" : [ "Solicitudes de combustible heredadas" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Servidor de la API",
    "url" : "http://localhost:8080"
  } ],
...
}